"Fix: Updated 'user_scorm_status' view to handle missing UserScormStatus object
This commit addresses an issue in the 'user_scorm_status' view where the system was not able to find a UserScormStatus object for a given SCORM ID. The view has been updated to create or update the UserScormStatus object if it does not exist. Additional logging has been added for better debugging and traceability. Also, the import error for 'datetime' and 'model_to_dict' has been resolved. Lastly, the response data has been customized to include only the necessary fields from the UserScormStatus object."